responded to a round of comments
