+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/annotator/ClippingRankSumTest.java
@@ -4,6 +4,7 @@ import htsjdk.variant.vcf.VCFInfoHeaderLine;
 import org.broadinstitute.barclay.help.DocumentedFeature;
 import org.broadinstitute.hellbender.utils.Utils;
 import org.broadinstitute.hellbender.utils.help.HelpConstants;
+import org.broadinstitute.hellbender.utils.pileup.PileupElement;
 import org.broadinstitute.hellbender.utils.read.AlignmentUtils;
 import org.broadinstitute.hellbender.utils.read.GATKRead;
 import org.broadinstitute.hellbender.utils.variant.GATKVCFConstants;
@@ -35,5 +36,12 @@ public final class ClippingRankSumTest extends RankSumTest implements StandardHC
     protected OptionalDouble getElementForRead(final GATKRead read, final int refLoc) {
         Utils.nonNull(read);
         return OptionalDouble.of(AlignmentUtils.getNumHardClippedBases(read));
+    }
+
+    @Override
+    protected OptionalDouble getElementForPileupElement(final PileupElement p, final int refLoc) {
+        Utils.nonNull(p);
+        // default to returning the same value
+        return getElementForRead(p.getRead(),refLoc);
     }
  }
