+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/annotator/ReadPosRankSumTest.java
@@ -1,17 +1,16 @@
 package org.broadinstitute.hellbender.tools.walkers.annotator;
 
-import htsjdk.variant.vcf.VCFInfoHeaderLine;
+import htsjdk.samtools.CigarElement;
+import htsjdk.samtools.CigarOperator;
 import org.broadinstitute.barclay.help.DocumentedFeature;
 import org.broadinstitute.hellbender.utils.Utils;
-import org.broadinstitute.hellbender.utils.pileup.PileupElement;
 import org.broadinstitute.hellbender.utils.help.HelpConstants;
+import org.broadinstitute.hellbender.utils.pileup.PileupElement;
 import org.broadinstitute.hellbender.utils.read.AlignmentUtils;
 import org.broadinstitute.hellbender.utils.read.GATKRead;
 import org.broadinstitute.hellbender.utils.read.ReadUtils;
 import org.broadinstitute.hellbender.utils.variant.GATKVCFConstants;
-import org.broadinstitute.hellbender.utils.variant.GATKVCFHeaderLines;
 
-import javax.swing.text.html.Option;
 import java.util.Collections;
 import java.util.List;
 import java.util.OptionalDouble;
@@ -34,6 +33,8 @@ import java.util.OptionalDouble;
  */
 @DocumentedFeature(groupName=HelpConstants.DOC_CAT_ANNOTATORS, groupSummary=HelpConstants.DOC_CAT_ANNOTATORS_SUMMARY, summary="Rank sum test for relative positioning of REF versus ALT alleles within reads (ReadPosRankSum)")
 public final class ReadPosRankSumTest extends RankSumTest implements StandardAnnotation {
+
+    public static final int LEGACY_PAIRHMM_BASE_QUALITY_SCORE_THRESHOLD = 20; //TODO the 20 hard value here is in lieu of PairHMM.BASE_QUALITY_SCORE_THRESHOLD in order to directly match GATK3 output
 
     @Override
     public List<String> getKeyNames() { return Collections.singletonList(GATKVCFConstants.READ_POS_RANK_SUM_KEY); }
@@ -74,6 +75,96 @@ public final class ReadPosRankSumTest extends RankSumTest implements StandardAnn
     @Override
     protected OptionalDouble getElementForPileupElement(final PileupElement p, int refLoc) {
         final int offset = AlignmentUtils.calcAlignmentByteArrayOffset(p.getRead().getCigar(), p, 0, 0);
-        return OptionalDouble.of(AnnotationUtils.getFinalVariantReadPosition(p.getRead(), offset));
+        return OptionalDouble.of(getFinalVariantReadPosition(p.getRead(), offset));
+    }
+
+    // Utility methods necessary for computing the rank sum using read pileups.
+    // TODO these methods contain bugs ported from GATK3, see issue https://github.com/broadinstitute/gatk/issues/4450 for progress towards fixing them
+
+    /**
+     * Get the position of a variant within a read with respect to the closer end, accounting for hard clipped bases and low quality ends
+     * Used by ReadPosRankSum annotations
+     *
+     * @param read  a read containing the variant
+     * @param initialReadPosition   the position based on the modified, post-hard-clipped CIGAR
+     * @return read position
+     */
+    public static int getFinalVariantReadPosition(final GATKRead read, final int initialReadPosition) {
+        final int numAlignedBases = getNumAlignedBases(read);
+
+        int readPos = initialReadPosition;
+        //TODO: this doesn't work for the middle-right position if we index from zero
+        if (initialReadPosition > numAlignedBases / 2) {
+            readPos = numAlignedBases - (initialReadPosition + 1);
+        }
+        return readPos;
+    }
+
+    /**
+     *
+     * @param read  read to count bases upon
+     * @return  the number of hard clipped and low qual bases at the read start (where start is the leftmost end w.r.t. the reference)
+     */
+    public static int getNumClippedBasesAtStart(final GATKRead read) {
+        // check for hard clips (never consider these bases):
+        final CigarElement first = read.getCigarElement(0);
+
+        final byte[] unclippedReadBases = read.getBases();
+        final byte[] unclippedReadQuals = read.getBaseQualities();
+
+        // Do a stricter base clipping than provided by CIGAR string, since this one may be too conservative,
+        // and may leave a string of Q2 bases still hanging off the reads.
+        //TODO: this code may not even get used because HaplotypeCaller already hard clips low quality tails
+
+        int numStartClippedBasesCounter = (first.getOperator() == CigarOperator.H)? first.getLength(): 0;
+
+        for (int i = numStartClippedBasesCounter; i < unclippedReadBases.length; i++) {
+            if (unclippedReadQuals[i] < LEGACY_PAIRHMM_BASE_QUALITY_SCORE_THRESHOLD) {
+                numStartClippedBasesCounter++;
+            } else {
+                break;
+            }
+        }
+
+        return numStartClippedBasesCounter;
+    }
+
+    /**
+     *
+     * @param read  a read containing the variant
+     * @return  number of non-hard clipped, aligned bases (excluding low quality bases at either end)
+     */
+    //TODO: this is bizarre -- this code counts hard clips, but then subtracts them from the read length, which already doesn't count hard clips
+    public static int getNumAlignedBases(final GATKRead read) {
+        return read.getLength() - getNumClippedBasesAtStart(read) - getNumClippedBasesAtEnd(read);
+    }
+
+    /**
+     *
+     * @param read  a read containing the variant
+     * @return  number of hard clipped and low qual bases at the read end (where end is right end w.r.t. the reference)
+     */
+    public static int getNumClippedBasesAtEnd(final GATKRead read) {
+        // check for hard clips (never consider these bases):
+        CigarElement last = read.getCigarElement(read.numCigarElements() - 1);
+
+        final byte[] unclippedReadBases = read.getBases();
+        final byte[] unclippedReadQuals = read.getBaseQualities();
+
+        // Do a stricter base clipping than provided by CIGAR string, since this one may be too conservative,
+        // and may leave a string of Q2 bases still hanging off the reads.
+        //TODO: this code may not even get used because HaplotypeCaller already hard clips low quality tails
+
+        int endClippedBasesCounter = last.getOperator() == CigarOperator.H ? last.getLength() : 0;
+
+        for (int i = unclippedReadBases.length - endClippedBasesCounter - 1; i >= 0; i--) {
+            if (unclippedReadQuals[i] < LEGACY_PAIRHMM_BASE_QUALITY_SCORE_THRESHOLD) {
+
+                endClippedBasesCounter++;
+            } else {
+                break;
+            }
+        }
+        return endClippedBasesCounter;
     }
 }
