--- src/main/java/org/broadinstitute/hellbender/tools/walkers/variantutils/StripAnnotations.java
@@ -1,4 +1,0 @@
-package org.broadinstitute.hellbender.tools.walkers.variantutils;
-
-public class StripAnnotations {
-}
